refactor: import type
